+++ lib/widgets/application_info_panel.dart
@@ -21,6 +21,7 @@ import 'package:flauncher/providers/apps_service.dart';
 import 'package:flauncher/widgets/right_panel_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class ApplicationInfoPanel extends StatelessWidget {
   final Category? category;
@@ -30,11 +31,14 @@ class ApplicationInfoPanel extends StatelessWidget {
   const ApplicationInfoPanel({
     required this.category,
     required this.application,
-    this.applicationIcon
+    required this.applicationIcon
   });
 
   @override
-  Widget build(BuildContext context) => RightPanelDialog(
+  Widget build(BuildContext context) {
+    AppLocalizations localizations = AppLocalizations.of(context)!;
+
+    return RightPanelDialog(
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
@@ -72,7 +76,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                 children: [
                   const Icon(Icons.open_in_new),
                   Container(width: 8),
-                  Text("Open", style: Theme.of(context).textTheme.bodyMedium),
+                  Text(localizations.open, style: Theme.of(context).textTheme.bodyMedium),
                 ],
               ),
               onPressed: () async {
@@ -86,7 +90,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                   children: [
                     const Icon(Icons.open_with),
                     Container(width: 8),
-                    Text("Reorder", style: Theme.of(context).textTheme.bodyMedium),
+                    Text(localizations.reorder, style: Theme.of(context).textTheme.bodyMedium),
                   ],
                 ),
                 onPressed: () => Navigator.of(context).pop(ApplicationInfoPanelResult.reorderApp),
@@ -96,7 +100,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                 children: [
                   Icon(application.hidden ? Icons.visibility : Icons.visibility_off_outlined),
                   Container(width: 8),
-                  Text(application.hidden ? "Unhide" : "Hide", style: Theme.of(context).textTheme.bodyMedium),
+                  Text(application.hidden ? localizations.show : localizations.hide, style: Theme.of(context).textTheme.bodyMedium),
                 ],
               ),
               onPressed: () async {
@@ -116,7 +120,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                     Container(width: 8),
                     Flexible(
                       child: Text(
-                        "Remove from ${category!.name}",
+                        localizations.removeFrom(category!.name),
                         style: Theme.of(context).textTheme.bodyMedium,
                         maxLines: 2,
                         overflow: TextOverflow.ellipsis,
@@ -135,7 +139,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                 children: [
                   const Icon(Icons.info_outlined),
                   Container(width: 8),
-                  Text("App info", style: Theme.of(context).textTheme.bodyMedium),
+                  Text(localizations.appInfo, style: Theme.of(context).textTheme.bodyMedium),
                 ],
               ),
               onPressed: () => context.read<AppsService>().openAppInfo(application),
@@ -145,7 +149,7 @@ class ApplicationInfoPanel extends StatelessWidget {
                 children: [
                   const Icon(Icons.delete_outlined),
                   Container(width: 8),
-                  Text("Uninstall", style: Theme.of(context).textTheme.bodyMedium),
+                  Text(localizations.uninstall, style: Theme.of(context).textTheme.bodyMedium),
                 ],
               ),
               onPressed: () async {
@@ -156,6 +160,7 @@ class ApplicationInfoPanel extends StatelessWidget {
           ],
         ),
       );
+  }
 }
 
 enum ApplicationInfoPanelResult { none, reorderApp }