+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Allow users to customize the date widget format, making it possible to display the [date, time](https://pub.dev/documentation/intl/latest/intl/DateFormat-class.html), and any arbitrary text, or to hide it completely.
 * Choose what the back button does while being in the launcher: Nothing, Show a fullscreen clock, Show screensaver (a.k.a Ambient Mode).
 * Show the active network status next to the date widget.
+* Allow users to hide the application category titles.
+* Clicking or tapping on an application card now opens it.
+* Translate to Spanish.
 
 ### Removals
 * Firebase analytics and crashlytics services.